Update assign_readers modules and test
