+++ gears/assign_readers/run.py
@@ -2,153 +2,24 @@
 
 import logging
 import os
-import re
-from pathlib import Path
 
-import pandas as pd
 from gear_toolkit import gear_toolkit_context
 
-from .utils.check_jobs import (
+from utils.check_jobs import (
     DuplicateJobError,
     InsufficientPermissionsError,
     check_for_duplicate_execution,
     verify_user_permissions,
 )
-from .utils.container_operations import find_or_create_group
-from .utils.manage_cases import (
+from utils.container_operations import find_or_create_group
+from utils.manage_cases import (
     InvalidGroupError,
     InvalidInputError,
     create_or_update_reader_projects,
+    define_reader_csv,
 )
 
 log = logging.getLogger(__name__)
-
-
-def define_reader_csv(context):
-    """
-    Loads, updates or creates a csv file based on gear input and configuration
-
-    If the reader_csv is specified in the gear configuration (and is valid) it is
-    loaded as a pandas dataframe.
-
-    If a specified reader is valid (email, firstname, lastname) it is appended
-    to the pandas dataframe (if invalid, skipped).
-
-    Without the reader_csv (or invalid) the specified reader is validated and saved to
-    a csv file in the context.work directory.  If specified reader is invalid,
-    None is returned.
-
-    Args:
-        context (gear_toolkit.GearContext): The gear context
-
-    Raises:
-        InvalidInputError: If neither the configuration (email, firstname, lastname) nor
-            the input (csv with fields email, firstname, lastname, max_cases) is valid
-            then this Error is thrown and the gear fails with message.
-
-    Returns:
-        str: The path of the resultant csv file or None (fail)
-    """
-    readers_df = []
-    # regex for checking validity of readers email
-    regex = r"^[a-zA-Z0-9.]+[\._]?[a-zA-Z0-9.]+[@]\w+[.]\w{2,3}$"
-    # Ensure valid inputs and act consistently
-    reader_csv_path = context.get_input_path("reader_csv")
-    if reader_csv_path:
-        readers_df = pd.read_csv(reader_csv_path)
-        # Validate that dataframe has required columns before proceeding
-        req_columns = ["email", "first_name", "last_name", "max_cases"]
-        if not all([(c in readers_df.columns) for c in req_columns]):
-            log.warning(
-                'The csv-file "%s" did not have the required columns("%s").'
-                + "Proceeding without reader CSV.",
-                Path(reader_csv_path).name,
-                '", "'.join(req_columns),
-            )
-            reader_csv_path = None
-        else:
-            # if we have a reader email, check for existence in csv (update),
-            # otherwise we need to create (if all conditions are satisfied)
-            if context.config.get("reader_email"):
-                reader_email = context.config.get("reader_email")
-                # if we find the reader's email in the dataframe,
-                if len(readers_df[readers_df.email == reader_email]) > 0:
-                    indx = readers_df[readers_df.email == reader_email].index[0]
-                    # Update the max_cases in the dataframe
-                    readers_df.loc[indx, "max_cases"] = context.config["max_cases"]
-                    # This will trigger an update in the metadata on assign-cases
-                # else if we have reader's email, firstname, and lastname
-                elif (
-                    context.config.get("max_cases")
-                    and (context.config.get("max_cases") > 0)
-                    and (context.config.get("max_cases") < 600)
-                    and context.config.get("reader_email")
-                    and re.search(regex, context.config.get("reader_email"))
-                    and context.config.get("reader_firstname")
-                    and context.config.get("reader_lastname")
-                ):
-                    readers_df = readers_df.append(
-                        {
-                            "email": context.config.get("reader_email"),
-                            "first_name": context.config.get("reader_firstname"),
-                            "last_name": context.config.get("reader_lastname"),
-                            "max_cases": context.config.get("max_cases"),
-                        },
-                        ignore_index=True,
-                    )
-                # else the indicated reader is invalid
-                else:
-                    log.warning(
-                        "The specified reader is not configured correctly. "
-                        'Proceeding without specified reader ("%s").',
-                        '", "'.join(
-                            [
-                                str(context.config.get("reader_email")),
-                                str(context.config.get("reader_firstname")),
-                                str(context.config.get("reader_lastname")),
-                            ]
-                        ),
-                    )
-
-            # Check the whole DataFrame for compliance to the regex on emails
-            if not all([re.search(regex, X) is not None for X in readers_df.email]):
-                raise InvalidInputError(
-                    "Cannot proceed without a valid CSV file or valid specified reader!"
-                )
-
-            # Create a csv and return its path
-            work_csv = context.work_dir / Path(reader_csv_path).name
-            readers_df.to_csv(work_csv, index=False)
-            return work_csv
-
-    # if the csv is not provided and we have a valid reader entry
-    if not reader_csv_path and (
-        context.config.get("max_cases")
-        and (context.config.get("max_cases") > 0)
-        and (context.config.get("max_cases") < 600)
-        and context.config.get("reader_email")
-        and re.search(regex, context.config.get("reader_email"))
-        and context.config.get("reader_firstname")
-        and context.config.get("reader_lastname")
-    ):
-        # create that dataframe
-        readers_df = pd.DataFrame(
-            data={
-                "email": context.config.get("reader_email"),
-                "first_name": context.config.get("reader_firstname"),
-                "last_name": context.config.get("reader_lastname"),
-                "max_cases": context.config.get("max_cases"),
-            },
-            index=[0],
-        )
-        # save it to the work directory
-        work_csv = context.work_dir / "temp.csv"
-        readers_df.to_csv(work_csv, index=False)
-        return work_csv
-    else:
-        raise InvalidInputError(
-            "Cannot proceed without a valid CSV file or valid specified reader!"
-        )
 
 
 def main(context):
